Test page with randomly assigned questions.
Created test.html
	Displays a list of questions.
	Needs GET to store the answers.

Created TestController.java
	Currently uses a placeholder user ID until we have implemented the way
of getting the current user.
	Currently has the default number of questions as 20. Did we plan on
setting this somewhere else?

Updated Test.java to include a list of questions, so that
springboot/thymeleaf can access them.
	Added setQuestions method to get all the questions from the repo with
the test's ID.

Updated Question.java to only get the noun in question and not its
translation when generateQuestion is called.